Adding Pubnub API
Added Pubnub API to project (added .jar's).   To acquire Pubnub API for
android, go to http://www.pubnub.com/docs/java/android/android-sdk.html
and "Download the entire Android SDK". Extract it to anywhere on your
drive.  When Android Studio is launched after doing this it will find
the *.class files for pubnub and you'll be good to go.